pyproject.toml
[tool.poetry]
name = "kr7"
version = "0.1.0"
description = ""
authors = ["Borodin Dmitriy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-multiselectfield = "^0.1.13"
eventlet = "^0.37.0"
redis = "^5.2.0"
coverage = "^7.6.6"
flake8 = "^7.1.1"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
